function that reverses a string

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -5,7 +5,7 @@
  **/
 void rev_string(char *s)
 {
-		short i, j;
+		int i, j;
 
 	for (i = 0; s[i] != '\0'; i++)
 	{}
